split datepicker, hour picker and vaccines into classes to be used in different activities
