+++ app/src/main/java/com/example/vaccinationapp/ui/schedule/ScheduleActivity.kt
@@ -28,6 +28,9 @@ import com.example.vaccinationapp.queries.AppointmentsQueries
 import com.example.vaccinationapp.queries.HealthcareUnitsQueries
 import com.example.vaccinationapp.queries.UsersQueries
 import com.example.vaccinationapp.queries.VaccinationsQueries
+import com.example.vaccinationapp.ui.Dates
+import com.example.vaccinationapp.ui.Hours
+import com.example.vaccinationapp.ui.Vaccines
 import com.example.vaccinationapp.ui.managerecords.ManageRecordsFragment
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.Dispatchers
@@ -53,6 +56,9 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
     private var dateTime = ""
     //FINAL VALUES
     private var FvaccineID: Int = 0
+    private val hoursManager = Hours()
+    private val vaccinesManager = Vaccines()
+    private val datesManager = Dates()
 
 
     @SuppressLint("MissingInflatedId", "SimpleDateFormat", "WeekBasedYear")
@@ -80,7 +86,7 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
             }
         }
 
-        val vaccines = offeredVaccines?.let { offerVaccines(it) }
+        val vaccines = offeredVaccines?.let { vaccinesManager.offerVaccines(it) }
 
         val vaccinesRecycler = findViewById<RecyclerView>(R.id.vaccinesRecycler)
         vaccinesRecycler.layoutManager = LinearLayoutManager(this)
@@ -113,13 +119,18 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
         }
 
 
-        val offeredHours = offerHours("08:00:00", "16:00:00", 30)
+        val offeredHours = hoursManager.offerHours("08:00:00", "16:00:00", 30)
         Log.d("OFFEREDHOURS", "$offeredHours")
 
         // PICKING THE DATE WILL RESULT IN THE APP SHOWING AVAILABLE HOURS
         date.setOnClickListener {
-            showDatePickerDialog(date, offeredHours)
+            val string = datesManager.showDatePickerDialog(this, date, offeredHours, hoursManager)
+            val stringSplit = string.split(";")
+            selectedDate = stringSplit[0]
+            selectedDateFormatted = stringSplit[1]
+            Log.d("DATES", "selected: $selectedDate, formatted: $selectedDateFormatted")
         }
+
 
         confirm.setOnClickListener {
             val splitDateTime = dateTime.split(";")
@@ -158,178 +169,178 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
 
     }
 
-    private fun offerVaccines(offeredVaccines:Set<Vaccinations>): List<String>{
-        val result = mutableListOf<String>()
-        var healthcareUnit: HealthcareUnits? = HealthcareUnits()
+//    private fun offerVaccines(offeredVaccines:Set<Vaccinations>): List<String>{
+//        val result = mutableListOf<String>()
+//        var healthcareUnit: HealthcareUnits? = HealthcareUnits()
+//
+//        if (offeredVaccines.isNotEmpty()){
+//            for (vaccine in offeredVaccines){
+//                val name = vaccine.vaccineName.toString()
+//                val id = vaccine.healthcareUnitId
+//                runBlocking { launch(Dispatchers.IO){
+//                    healthcareUnit = id?.let { getHealthcareUnit(it) }
+//                } }
+//                val unitName = healthcareUnit?.name
+//                val city = healthcareUnit?.city
+//                val street = healthcareUnit?.street
+//                val nr = healthcareUnit?.streetNumber
+//                val address = "$unitName $city $street $nr"
+//                val nameAddress = "$name;$address;$id"
+//                result.add(nameAddress)
+//            }
+//        }
+//        return result
+//    }
 
-        if (offeredVaccines.isNotEmpty()){
-            for (vaccine in offeredVaccines){
-                val name = vaccine.vaccineName.toString()
-                val id = vaccine.healthcareUnitId
-                runBlocking { launch(Dispatchers.IO){
-                    healthcareUnit = id?.let { getHealthcareUnit(it) }
-                } }
-                val unitName = healthcareUnit?.name
-                val city = healthcareUnit?.city
-                val street = healthcareUnit?.street
-                val nr = healthcareUnit?.streetNumber
-                val address = "$unitName $city $street $nr"
-                val nameAddress = "$name;$address;$id"
-                result.add(nameAddress)
-            }
-        }
-        return result
-    }
+//    private fun showDatePickerDialog(date: Button, offeredHours: List<String>): String {
+//        val calendar = Calendar.getInstance()
+//        val currentYear = calendar.get(Calendar.YEAR)
+//        val currentMonth = calendar.get(Calendar.MONTH)
+//        val currentDay = calendar.get(Calendar.DAY_OF_MONTH)
+//
+//        val datePickerDialog = DatePickerDialog(
+//            this,
+//            { _: DatePicker?, selectedYear: Int, selectedMonth: Int, selectedDay: Int ->
+//                selectedDate = "$selectedYear-${selectedMonth+1}-$selectedDay"
+//                val monthName = DateFormatSymbols().months[selectedMonth]
+//                selectedDateFormatted = "$selectedDay $monthName $selectedYear"
+//
+//                if (isDateValid(selectedYear, selectedMonth, selectedDay)) {
+//                    date.text = selectedDateFormatted
+//
+//                    //check available hours for picked date
+//                    val availableHours = hoursManager.getAvailableHours(selectedDate, offeredHours)
+//
+//                    val timeRecycler = findViewById<RecyclerView>(R.id.hoursRecycler)
+//                    timeRecycler.layoutManager = LinearLayoutManager(this)
+//
+//                    val adapterTime = HoursAdapter(availableHours, date)
+//                    timeRecycler.adapter = adapterTime
+//
+//                    adapterTime.setOnItemClickListener(this)
+//                }
+//            },
+//            currentYear,
+//            currentMonth,
+//            currentDay
+//        )
+//
+//        datePickerDialog.datePicker.minDate = System.currentTimeMillis()
+//        datePickerDialog.show()
+//
+//        return selectedDate
+//    }
 
-    private fun showDatePickerDialog(date: Button, offeredHours: List<String>): String {
-        val calendar = Calendar.getInstance()
-        val currentYear = calendar.get(Calendar.YEAR)
-        val currentMonth = calendar.get(Calendar.MONTH)
-        val currentDay = calendar.get(Calendar.DAY_OF_MONTH)
+//    @SuppressLint("SimpleDateFormat")
+//    private fun getAvailableHours(selectedDate: String, offeredHours: List<String>): List<String>{
+//        var takenHours: List<String>? = null
+//        var offeredHoursFormatted: List<String>? = null
+//        try {
+//            runBlocking {
+//                launch(Dispatchers.IO) {
+//                    takenHours = getAllAppointmentsForDate(selectedDate)
+//                }
+//            }
+//            Log.d("TAKENHOURS", "$takenHours")
+//
+//        } catch (e: Exception) {
+//            e.printStackTrace()
+//        }
+//
+//        takenHours = formatHours(takenHours)
+//        offeredHoursFormatted = formatHours(offeredHours)
+//
+////        // get current date and time
+//        val calendar = Calendar.getInstance()
+//        calendar.timeZone = TimeZone.getTimeZone("CET")
+//        val currentDate = "${calendar.get(Calendar.YEAR)}-${calendar.get(Calendar.MONTH)+1}-${calendar.get(Calendar.DAY_OF_MONTH)}"
+//        val currentTime= "${calendar.get(Calendar.HOUR_OF_DAY)}:${calendar.get((Calendar.MINUTE))}"
+//        val sdf = SimpleDateFormat("H:mm")
+//
+//        if(selectedDate == currentDate) {
+//            offeredHoursFormatted = filterHours(offeredHoursFormatted, currentTime)
+//        }
+//
+//        return if (takenHours == null) {
+//            offeredHoursFormatted
+//        }else{
+//            val availableHours = offeredHoursFormatted.minus(takenHours ?: emptyList())
+//            availableHours
+//        }
+//    }
+//
+//    fun filterHours(offeredHours: List<String>, currentTime: String): List<String> {
+//        // Parse the current time
+//        val currentTimeSplit = currentTime.split(":")
+//        val currentHour = currentTimeSplit[0].toInt()
+//        val currentMinute = currentTimeSplit[1].toInt()
+//
+//        // Function to parse a time string and return the hour and minute as a pair
+//        fun parseTime(time: String): Pair<Int, Int> {
+//            val parts = time.split(":")
+//            val hour = parts[0].toInt()
+//            val minute = parts[1].toInt()
+//            return Pair(hour, minute)
+//        }
+//
+//        // Filter the times based on the comparison with the current time
+//        return offeredHours.filter { time ->
+//            val (hour, minute) = parseTime(time)
+//            if (hour < currentHour) {
+//                false
+//            } else if (hour == currentHour) {
+//                minute >= currentMinute
+//            } else {
+//                true
+//            }
+//        }
+//    }
+//
+//    fun formatHours(hours: List<String>?): List<String> {
+//        return hours?.map { time ->
+//            val parts = time.split(":")
+//            val formattedHour = parts[0].trimStart('0')
+//            "$formattedHour:${parts[1]}"
+//        } ?: emptyList()
+//    }
+//
+//    //code that generates offered hours (admin can choose the working hours they offer to patients
+//    //as well as choosing the time interval between the visits)
+//    fun offerHours(startTime: String, endTime: String, timeSkip: Int): List<String>{
+//        val offeredHours = mutableListOf<String>()
+//        var currentTime = startTime
+//
+//        while (currentTime != endTime){
+//            offeredHours.add(currentTime)
+//            currentTime = addHalfHour(currentTime, timeSkip)
+//        }
+//        offeredHours.add(endTime)
+//        return offeredHours
+//    }
+//    private fun addHalfHour(currentTime: String, timeSkip: Int): String{
+//        val (hour, minute) = currentTime.split(":").map{it.toInt()}
+//        // calculate next hour in minutes (+30 for next available hour)
+//        val totalMinutes = hour * 60 + minute + timeSkip
+//        val nextHour = totalMinutes/60
+//        val nextMinute = totalMinutes%60
+//        return String.format("%02d:%02d:00", nextHour, nextMinute) // :00 !!!!!!!!!!!!!!!!!!!!!!!!
+//    }
 
-        val datePickerDialog = DatePickerDialog(
-            this,
-            { _: DatePicker?, selectedYear: Int, selectedMonth: Int, selectedDay: Int ->
-                selectedDate = "$selectedYear-${selectedMonth+1}-$selectedDay"
-                val monthName = DateFormatSymbols().months[selectedMonth]
-                selectedDateFormatted = "$selectedDay $monthName $selectedYear"
-
-                if (isDateValid(selectedYear, selectedMonth, selectedDay)) {
-                    date.text = selectedDateFormatted
-
-                    //check available hours for picked date
-                    val availableHours = getAvailableHours(selectedDate, offeredHours)
-
-                    val timeRecycler = findViewById<RecyclerView>(R.id.hoursRecycler)
-                    timeRecycler.layoutManager = LinearLayoutManager(this)
-
-                    val adapterTime = HoursAdapter(availableHours, date)
-                    timeRecycler.adapter = adapterTime
-
-                    adapterTime.setOnItemClickListener(this)
-                }
-            },
-            currentYear,
-            currentMonth,
-            currentDay
-        )
-
-        datePickerDialog.datePicker.minDate = System.currentTimeMillis()
-        datePickerDialog.show()
-
-        return selectedDate
-    }
-
-    @SuppressLint("SimpleDateFormat")
-    private fun getAvailableHours(selectedDate: String, offeredHours: List<String>): List<String>{
-        var takenHours: List<String>? = null
-        var offeredHoursFormatted: List<String>? = null
-        try {
-            runBlocking {
-                launch(Dispatchers.IO) {
-                    takenHours = getAllAppointmentsForDate(selectedDate)
-                }
-            }
-            Log.d("TAKENHOURS", "$takenHours")
-
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-        takenHours = formatHours(takenHours)
-        offeredHoursFormatted = formatHours(offeredHours)
-
-//        // get current date and time
-        val calendar = Calendar.getInstance()
-        calendar.timeZone = TimeZone.getTimeZone("CET")
-        val currentDate = "${calendar.get(Calendar.YEAR)}-${calendar.get(Calendar.MONTH)+1}-${calendar.get(Calendar.DAY_OF_MONTH)}"
-        val currentTime= "${calendar.get(Calendar.HOUR_OF_DAY)}:${calendar.get((Calendar.MINUTE))}"
-        val sdf = SimpleDateFormat("H:mm")
-
-        if(selectedDate == currentDate) {
-            offeredHoursFormatted = filterHours(offeredHoursFormatted, currentTime)
-        }
-
-        return if (takenHours == null) {
-            offeredHoursFormatted
-        }else{
-            val availableHours = offeredHoursFormatted.minus(takenHours ?: emptyList())
-            availableHours
-        }
-    }
-
-    fun filterHours(offeredHours: List<String>, currentTime: String): List<String> {
-        // Parse the current time
-        val currentTimeSplit = currentTime.split(":")
-        val currentHour = currentTimeSplit[0].toInt()
-        val currentMinute = currentTimeSplit[1].toInt()
-
-        // Function to parse a time string and return the hour and minute as a pair
-        fun parseTime(time: String): Pair<Int, Int> {
-            val parts = time.split(":")
-            val hour = parts[0].toInt()
-            val minute = parts[1].toInt()
-            return Pair(hour, minute)
-        }
-
-        // Filter the times based on the comparison with the current time
-        return offeredHours.filter { time ->
-            val (hour, minute) = parseTime(time)
-            if (hour < currentHour) {
-                false
-            } else if (hour == currentHour) {
-                minute >= currentMinute
-            } else {
-                true
-            }
-        }
-    }
-
-    fun formatHours(hours: List<String>?): List<String> {
-        return hours?.map { time ->
-            val parts = time.split(":")
-            val formattedHour = parts[0].trimStart('0')
-            "$formattedHour:${parts[1]}"
-        } ?: emptyList()
-    }
-
-    //code that generates offered hours (admin can choose the working hours they offer to patients
-    //as well as choosing the time interval between the visits)
-    fun offerHours(startTime: String, endTime: String, timeSkip: Int): List<String>{
-        val offeredHours = mutableListOf<String>()
-        var currentTime = startTime
-
-        while (currentTime != endTime){
-            offeredHours.add(currentTime)
-            currentTime = addHalfHour(currentTime, timeSkip)
-        }
-        offeredHours.add(endTime)
-        return offeredHours
-    }
-    private fun addHalfHour(currentTime: String, timeSkip: Int): String{
-        val (hour, minute) = currentTime.split(":").map{it.toInt()}
-        // calculate next hour in minutes (+30 for next available hour)
-        val totalMinutes = hour * 60 + minute + timeSkip
-        val nextHour = totalMinutes/60
-        val nextMinute = totalMinutes%60
-        return String.format("%02d:%02d:00", nextHour, nextMinute) // :00 !!!!!!!!!!!!!!!!!!!!!!!!
-    }
-
-    fun isDateValid(selectedYear: Int, selectedMonth: Int, selectedDay: Int): Boolean {
-        // Check if the year, month, and day are within valid ranges
-        if (selectedYear < 0 || selectedMonth < 1 || selectedMonth > 12 || selectedDay < 1) {
-            return false
-        }
-
-        // Check if the day is within the valid range for the selected month
-        val daysInMonth = when (selectedMonth) {
-            1, 3, 5, 7, 8, 10, 12 -> 31
-            4, 6, 9, 11 -> 30
-            2 -> if (selectedYear % 4 == 0 && (selectedYear % 100 != 0 || selectedYear % 400 == 0)) 29 else 28
-            else -> return false
-        }
-        return selectedDay <= daysInMonth
-    }
+//    fun isDateValid(selectedYear: Int, selectedMonth: Int, selectedDay: Int): Boolean {
+//        // Check if the year, month, and day are within valid ranges
+//        if (selectedYear < 0 || selectedMonth < 1 || selectedMonth > 12 || selectedDay < 1) {
+//            return false
+//        }
+//
+//        // Check if the day is within the valid range for the selected month
+//        val daysInMonth = when (selectedMonth) {
+//            1, 3, 5, 7, 8, 10, 12 -> 31
+//            4, 6, 9, 11 -> 30
+//            2 -> if (selectedYear % 4 == 0 && (selectedYear % 100 != 0 || selectedYear % 400 == 0)) 29 else 28
+//            else -> return false
+//        }
+//        return selectedDay <= daysInMonth
+//    }
 
     override fun onHourClick(item: String, date: Button) {
         val finalDate = "$selectedDateFormatted $item"
@@ -342,17 +353,17 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
     }
 
     //QUERIES
-    private suspend fun getAllAppointmentsForDate(date:String): List<String>?{
-        return withContext(Dispatchers.IO){
-            val connection = DBconnection.getConnection()
-            Log.d("DATABASE", "connected with date: $date")
-            val appointmentQueries = AppointmentsQueries(connection)
-            val result = appointmentQueries.getAllAppointmentsForDate(date)
-            Log.d("DATABASE", "result: $result")
-            connection.close()
-            result
-        }
-    }
+//    private suspend fun getAllAppointmentsForDate(date:String): List<String>?{
+//        return withContext(Dispatchers.IO){
+//            val connection = DBconnection.getConnection()
+//            Log.d("DATABASE", "connected with date: $date")
+//            val appointmentQueries = AppointmentsQueries(connection)
+//            val result = appointmentQueries.getAllAppointmentsForDate(date)
+//            Log.d("DATABASE", "result: $result")
+//            connection.close()
+//            result
+//        }
+//    }
 
     private suspend fun addAppointment(appointment: Appointments):Boolean{
         return withContext(Dispatchers.IO){
@@ -378,16 +389,16 @@ class ScheduleActivity : AppCompatActivity(), HoursAdapter.OnItemClickListener,
         }
     }
 
-    private suspend fun getHealthcareUnit(id: Int): HealthcareUnits?{
-        return withContext(Dispatchers.IO){
-            val conn = DBconnection.getConnection()
-            val HUqueries = HealthcareUnitsQueries(conn)
-            val result = HUqueries.getHealthcareUnit(id)
-            Log.d("DATABASE", "healthcare units: $result")
-            conn.close()
-            result
-        }
-    }
+//    private suspend fun getHealthcareUnit(id: Int): HealthcareUnits?{
+//        return withContext(Dispatchers.IO){
+//            val conn = DBconnection.getConnection()
+//            val HUqueries = HealthcareUnitsQueries(conn)
+//            val result = HUqueries.getHealthcareUnit(id)
+//            Log.d("DATABASE", "healthcare units: $result")
+//            conn.close()
+//            result
+//        }
+//    }
 
     private suspend fun getUserId(email: String): Int? {
         return withContext(Dispatchers.IO){
