+++ app/src/main/res/layout/activity_schedule.xml
@@ -16,7 +16,9 @@
         android:layout_marginTop="20dp"
         android:layout_marginEnd="30dp"
         android:backgroundTint="#5FBFBC"
+        android:fontFamily="@font/neuropol_x_rg"
         android:text="select date"
+        android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -50,14 +52,15 @@
 
     <Button
         android:id="@+id/confirmButton"
-        android:layout_width="110dp"
-        android:layout_height="50dp"
+        android:layout_width="146dp"
+        android:layout_height="49dp"
         android:layout_marginBottom="25dp"
         android:backgroundTint="#5FBEBB"
+        android:fontFamily="@font/neuropol_x_rg"
         android:text="confirm"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.75"
+        app:layout_constraintHorizontal_bias="0.811"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/hoursRecycler" />
 
@@ -75,16 +78,19 @@
 
     <Button
         android:id="@+id/cancelButton"
-        android:layout_width="110dp"
-        android:layout_height="50dp"
+        android:layout_width="130dp"
+        android:layout_height="47dp"
         android:layout_marginStart="90dp"
         android:layout_marginEnd="26dp"
         android:layout_marginBottom="25dp"
         android:backgroundTint="#5FBEBB"
+        android:fontFamily="@font/neuropol_x_rg"
         android:text="cancel"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/confirmButton"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/hoursRecycler" />
+        app:layout_constraintTop_toBottomOf="@+id/hoursRecycler"
+        app:layout_constraintVertical_bias="0.475" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
