+++ app/src/main/res/layout/fragment_manage_records.xml
@@ -42,7 +42,7 @@
         android:layout_marginTop="20dp"
         android:backgroundTint="#FF4FAEAF"
         android:fontFamily="@font/neuropol_x_rg"
-        android:text="add"
+        android:text="schedule"
         android:textSize="24sp"
         app:cornerRadius="10dp"
         app:layout_constraintEnd_toEndOf="parent"
